# api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Sistema de Gestion de Documentos y Procedimientos Administrativos Electrónicos"
  version: "1.0.0"
  title: "Mara API"
host: "localhost:8080"
...
basePath: "/v1"
schemes:
- "http"
produces:
- "application/json"
paths:
...
  /tods/{id}:
    get:
      tags:
      - "Tipo Documental"
      summary: "Tipo de objeto documental por id"
      description: "Obtiene un tipo de objeto documental segun el id"
      operationId: "todIdGet"
      parameters:
      - name: "id"
        in: "path"
        description: "Identificador único del tipo de objeto documental."
        required: true
        type: "string"
      responses:
        200:
          description: Tipo de objeto documental según el id provisto"
          schema:
            $ref: "#/definitions/TOD"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"

    delete:
      tags:
      - "Tipo Documental"
      summary: "Eliminación de tipo de objeto documental por id"
      description: "Eliminación de un tipo de objeto documental segun el id"
      operationId: "todIdDelete"
      parameters:
      - name: "id"
        in: "path"
        description: "Identificador del tipo de objeto documental."
        required: true
        type: "string"
      responses:
        204:
          description: "Tipo de objeto documental borrado."
          schema:
            $ref: "#/definitions/ReturnId"
        400:
          description: "Id incorrecto"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Recurso no encontrado"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Error interno del servidor"
          schema:
            $ref: "#/definitions/Error"

    x-swagger-router-controller: "tod"


# ------------------ DOCUMENTOS ------------------
...
definitions:
  TOD:
    type: "object"
    required:
      - label
      - icon
    properties:
      id:
        type: "string"
        description: "Identificador Unico."
      type:
        type: "string"
        description: "Define si es un tipo documental Document o Folder"
      label:
        type: "string"
        description: "Nombre completo del tipo de objeto documental."
      identifier:
        type: "string"
        description: "Identificador alfabetico del tipo de objeto documental utilizado para conformar el CUDAP."
      icon:
        type: "string"
        # format: "byte"
        description: "Icono del tipo de objeto documental. Tamaño 16 px"
      large-icon:
        type: "string"
        # format: "byte"
        description: "Icono del tipo de objeto documental. Tamaño 64 px"
      description:
        type: "string"
        description: "Descripcion del tipo de objeto documental."
      file-format:
        type: array
        description: "Listado de formato permitido de archivos adjuntos."
        items:
          type: "string"
      schemas:
        type: array
        description: "Listado de metadatos del tipo de objeto documental."
        items:
          type: "string"
      container-type:
        type: array
        description: "Listado de tipos de objetos documentales que puede contener si es un Folder."
        items:
          type: "string"
      # view-url:
      #   type: "string"
      #   description: "Url de la vista de un tipo de objeto documental."
      # edit-url:
      #   type: "string"
      #   description: "Url de edición de datos de la instancia del tipo de objeto documental."
      # new-url:
      #   type: "string"
      #   description: "Url de creación de una instancia del tipo de objeto documental"
      # links:
      #   type: "array"
      #   description: "Listado de vínculos a recursos"
      #   items:
      #     $ref: "#/definitions/Hypermedia"
...
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"
  ReturnId:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      id:
        type: "string"
  Hypermedia:
    type: "object"
    properties:
      rel:
        type: "string"
        description: "Nombre del recurso con el que se relaciona"
      href:
        type: "string"
        description: "URL de referencia al recurso"